+++ Packages/com.unity.inputsystem/InputSystem/Controls/InputControl.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 using System.Diagnostics;
 using UnityEngine.Experimental.Input.Controls;
 using UnityEngine.Experimental.Input.LowLevel;
@@ -273,7 +274,7 @@ namespace UnityEngine.Experimental.Input
         /// gamepad["{PrimaryAction}"] // Returns the control with PrimaryAction usage, i.e. Gamepad.aButton
         /// </code>
         /// </example>
-        /// <exception cref="IndexOutOfRangeException"><paramref name="path"/> cannot be found.</exception>
+        /// <exception cref="KeyNotFoundException"><paramref name="path"/> cannot be found.</exception>
         /// <seealso cref="InputControlPath"/>
         /// <seealso cref="path"/>
         /// <seealso cref="TryGetChildControl"/>
@@ -283,7 +284,7 @@ namespace UnityEngine.Experimental.Input
             {
                 var control = InputControlPath.TryFindChild(this, path);
                 if (control == null)
-                    throw new IndexOutOfRangeException(
+                    throw new KeyNotFoundException(
                         $"Cannot find control '{path}' as child of '{this}'");
                 return control;
             }