+++ Packages/com.unity.inputsystem/Tests/InputSystem/CoreTests_Actions.cs
@@ -2598,12 +2598,12 @@ partial class CoreTests
         float? receivedFloat = null;
         action.performed +=
             ctx =>
-            {
-                Assert.That(receivedFloat, Is.Null);
-                // ConstantVector2TestProcessor processes Vector2s. It would throw an exception when 
-                // trying to use it reading a float if not ignored.
-                receivedFloat = ctx.ReadValue<float>();
-            };
+        {
+            Assert.That(receivedFloat, Is.Null);
+            // ConstantVector2TestProcessor processes Vector2s. It would throw an exception when
+            // trying to use it reading a float if not ignored.
+            receivedFloat = ctx.ReadValue<float>();
+        };
 
         Set(gamepad.leftStick, Vector2.one);
 
@@ -3606,13 +3606,11 @@ partial class CoreTests
         Assert.That(asset.FindAction($"{{{action3.id.ToString()}}}"), Is.SameAs(action3));
 
         // Shouldn't allocate.
-        #if UNITY_2018_3_OR_NEWER
         var map1action1 = "map1/action1";
         Assert.That(() =>
         {
             asset.FindAction(map1action1);
         }, Is.Not.AllocatingGCMemory());
-        #endif
     }
 
     [Test]
@@ -3959,7 +3957,6 @@ partial class CoreTests
         }
     }
 
-    #if UNITY_2018_3_OR_NEWER
     [Test]
     [Category("Actions")]
     [Ignore("TODO")]
@@ -3984,8 +3981,6 @@ partial class CoreTests
             }, Is.Not.AllocatingGCMemory());
         }
     }
-
-    #endif
 
     [Test]
     [Category("Actions")]